refactor RCT1 import code
